Adding urlconf to uploads app.
Applications are meant to be reusable.  In the case of url configuration, the
urls used in a particular application can be placed in their own urlconf
module.  They can be included by other urlconfs by using the include()
function.

Because we are adding URLs that refer to non-existant views, stub them in for
the time being.