Creating a list view for uploaded files.
In this particular project I'm keeping the base template for the site in the
webapp directory.  By default, django looks for templates any directory named
"templates" within the installed applications, but it does not look for a
templates directory in the project.  We can add this in the settings.py file.

Application-specific templates are placed in the uploads application.