Adding login_required decorator.
Aside from the get view, all other views are user-specific, so make sure a
user is logged in.

The login_required decorator will make sure the request is being made by an
authenticated user; if not it will redirect to the login page, which is
/accounts/login/ by default.  This is customizable in settings.py.

In the webapp's urlconf I'm making use of the default login view and overriding
the logout view.