Adding upload form.
Django provides a form processing mechanism where you can define a form and it
takes care of validation and saving the object to the database.

Here we are creating a Form object based on the File model.  Since we don't
want to give users the ability to select what user to associate this file with
(instead it will automatically be associated with the logged in user when it's
uploaded) we can specify what fields from the model should be seen in the form.