Adding view to get file.
The get view simply redirects to the actual file being requested.

Django is not meant to serve static files, but it can be configured to do so
for development.  This is a really bad thing to do in production, so this
project makes sure that it will only be used when in debug mode.